edits components and sets up router/routes/route for react router
