+++ client/src/App.jsx
@@ -1,37 +1,35 @@
 import { useState, useEffect } from 'react';
+//'switch' has been replaced with 'routes'
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+
+import CreateRoseBudThorn from './components/CreateRoseBudThorn';
+import HomePage from './components/HomePage';
+import ListLatestEntrie from './components/ListLatestEntries';
+import RequestBouquet from './components/RequestBouquet';
 
 import './App.css'
 
 function App() {
 
-  const [backEndMessage, setBackEndMessage] = useState("");
-
-  const callBackEnd = async () => {
-    try{
-      const response = await fetch('http://localhost:8888');
-      if(!response.ok) {
-        throw new Error(`Request failed: ${response.status}`);
-      }
-      const data = await response.json();
-      setBackEndMessage(data);
-    } catch{
-      console.error('An error occured:', error);
-    }
-  }
-
-  useEffect(() => {
-    callBackEnd();
-  }, []);
-
 
   return (
-    <>
-      <div>
-        <h1>rose: a therapy companion</h1>
-        <p>{backEndMessage}</p>
-        <button>add a rose</button>
-      </div>
-    </>
+
+      <Router>
+
+        <div className='App'>
+
+          {/* navbar? */}
+
+          <div className='content'>
+            <Routes>
+              <Route path="/" element={<HomePage />} />
+              <Route path="/create" element={<CreateRoseBudThorn />} />
+            </Routes>
+          </div>
+
+        </div>
+
+      </Router>
   )
 }
