+++ client/src/components/CreateRoseBudThorn.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+export default function CreateRoseBudThorn() {
+  return (
+    <div classname="create" >
+
+    <h2>Create a new RoseBudThorn</h2>
+
+    </div>
+  )
+}
+
+//sfc tab
