+++ client/src/components/HomePage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function HomePage() {
+  return (
+    <div>
+        <h1>rose: a therapy companion</h1>
+    </div>
+  )
+}
