+++ client/src/components/ListLatestEntries.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function ListLatestEntries() {
+  return (
+    <div>ListLatestEntries</div>
+  )
+}
