+++ client/src/components/RequestBouquet.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function RequestBouquet() {
+  return (
+    <div>RequestBouquet</div>
+  )
+}
